room-monitoring-serial-comm/requirements.txt
pyserial==3.4
requests==2.21.0